+++ core/src/main/kotlin/no/elg/hex/event/EventListeners.kt
@@ -1,9 +1,8 @@
 package no.elg.hex.event
 
-import com.badlogic.gdx.utils.Array
 import no.elg.hex.event.events.Event
 
-open class EventListeners<T : Event>(val listeners: Array<(T) -> Unit> = Array<(T) -> Unit>())
+open class EventListeners<T : Event>(val listeners: MutableList<(T) -> Unit> = mutableListOf())
 
 @Retention(AnnotationRetention.RUNTIME)
 annotation class ClearOnScreenChange